@@ -1,6 +1,6 @@
 {
   "name": "simple-react-form",
-  "version": "1.9.4",
+  "version": "2.0.0",
   "license": "MIT",
   "author": "Nicolás López <nicolaslopezj@me.com>",
   "description": "A library to make reusable form components in React and React Native",
@@ -15,7 +15,7 @@
     "test": "yarn run build && yarn run lint && jest",
     "watch": "node watch.js",
     "test-watch": "jest --watch",
-    "lint": "standard \"src/**/*.js\""
+    "lint": "eslint \"src/**/*.js\""
   },
   "keywords": [
     "react-component",
@@ -26,25 +26,29 @@
     "react"
   ],
   "dependencies": {
-    "dot-object": "^1.5.1",
-    "lodash": "^4.17.2"
+    "autobind-decorator": "^2.1.0",
+    "lodash": "^4.17.4"
   },
   "devDependencies": {
-    "babel-cli": "6.5.1",
-    "babel-eslint": "^5.0.0",
-    "babel-jest": "^15.0.0",
+    "babel-cli": "^6.24.1",
+    "babel-eslint": "^7.2.3",
+    "babel-jest": "^20.0.3",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "6.5.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-2": "^6.5.0",
-    "enzyme": "^2.4.1",
-    "eslint": "^1.10.3",
-    "eslint-plugin-react": "^3.16.1",
-    "jest-cli": "^15.1.1",
-    "react": ">=15.0.0",
-    "react-addons-test-utils": "^15.3.2",
-    "react-dom": ">=15.0.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "colors": "^1.1.2",
+    "enzyme": "^2.9.1",
+    "eslint": "^4.4.1",
+    "eslint-config-orionsoft": "^1.2.1",
+    "eslint-plugin-babel": "^4.1.2",
+    "eslint-plugin-react": "^7.1.0",
+    "jest-cli": "^20.0.4",
+    "react": "^15.6.1",
+    "react-addons-test-utils": "^15.6.0",
+    "react-dom": "^15.6.1",
+    "react-test-renderer": "^15.6.1",
     "standard": "*"
   },
   "peerDependencies": {
