@@ -1,11 +1,5 @@
 export {default as Form} from './Form/index'
 export {default as Field} from './Field'
 export {default as FieldType} from './FieldType'
-export {default as ArrayComponent} from './Array/index'
+export {default as ArrayComponent} from './Array'
 export {default as ObjectComponent} from './Object'
-export {
-  registerType,
-  getFieldType,
-  getFieldTypeName,
-  getFieldComponent
-} from './types'
